requirements.txt
asgiref==3.3.1
django==3.1.5
djangorestframework==3.12.2
joblib==1.0.0
numpy==1.19.5
pandas==1.2.1
pickle-mixin==1.0.2
python-dateutil==2.8.1
pytz==2020.5
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
